+++ app/src/main/java/com/example/togedy_android/domain/model/planner/PlannerHomeInformation.kt
@@ -1,0 +1,7 @@
+package com.example.togedy_android.domain.model.planner
+
+data class PlannerHomeInformation(
+    val studyGoal: StudyGoal,
+    val studyPlanList: List<StudyPlanItem> = emptyList<StudyPlanItem>(),
+    val studyTagItemLists: List<StudyTagItem> = emptyList<StudyTagItem>()
+)
