+++ app/src/main/java/com/example/togedy_android/presentation/planner/planner/state/PlannerUiState.kt
@@ -1,7 +1,7 @@
 package com.example.togedy_android.presentation.planner.planner.state
 
 import com.example.togedy_android.core.state.UiState
-import com.example.togedy_android.domain.entity.PlannerHomeInformation
+import com.example.togedy_android.domain.model.planner.PlannerHomeInformation
 import com.example.togedy_android.domain.model.planner.StudyGoal
 import java.time.LocalDate
